moved extension.cs stuff to GeneralHelper
